fix: fix assistant layout
