+++ packages/frontend/editor-ui/src/app/layouts/WorkflowLayout.vue
@@ -1,9 +1,16 @@
 <script lang="ts" setup>
-import { defineAsyncComponent } from 'vue';
+import { ref, defineAsyncComponent } from 'vue';
 import BaseLayout from './BaseLayout.vue';
 import { useLayoutProps } from '@/app/composables/useLayoutProps';
+import AskAssistantFloatingButton from '@/features/ai/assistant/components/Chat/AskAssistantFloatingButton.vue';
+import AppChatPanel from '@/app/components/app/AppChatPanel.vue';
+import { useAssistantStore } from '@/features/ai/assistant/assistant.store';
 
 const { layoutProps } = useLayoutProps();
+
+const assistantStore = useAssistantStore();
+
+const layoutRef = ref<Element | null>(null);
 
 const AppHeader = defineAsyncComponent(
 	async () => await import('@/app/components/app/AppHeader.vue'),
@@ -14,10 +21,14 @@ const AppSidebar = defineAsyncComponent(
 const LogsPanel = defineAsyncComponent(
 	async () => await import('@/features/execution/logs/components/LogsPanel.vue'),
 );
+
+const setLayoutRef = (el: Element) => {
+	layoutRef.value = el;
+};
 </script>
 
 <template>
-	<BaseLayout>
+	<BaseLayout @mounted="setLayoutRef">
 		<template #header>
 			<Suspense>
 				<AppHeader />
@@ -33,6 +44,10 @@ const LogsPanel = defineAsyncComponent(
 				<Component :is="Component" />
 			</KeepAlive>
 		</RouterView>
+		<AskAssistantFloatingButton v-if="assistantStore.isFloatingButtonShown" />
+		<template #aside>
+			<AppChatPanel v-if="layoutRef" :layout-ref="layoutRef" />
+		</template>
 		<template v-if="layoutProps.logs" #footer>
 			<Suspense>
 				<LogsPanel />
